requirements.txt
Flask==0.12.2
line-bot-sdk==1.5.0
gunicorn==19.9.0
requests==2.22.0
googletrans==2.4.0
numpy==1.18.2
requests-oauthlib==1.3.0
Janome==0.3.10
wordcloud==1.7.0
Pillow==7.2.0
